refactor: declare filter types
- Remove type `FilterType`
- Create the type `FilterParams`. The `RpcFilterRequest` has some
  union-style fields, such as *address*, which can be null, a single
  address, or multiple addresses. `FilterParams` can be more specific
  and explicitly for use.
